Add proper unload logic
Before it was possible that race-conditions would cause tests to fail
and a possible unclean unload in the server could cause memory leaks.

This uses more done callbacks to ensure that unload and close operations
in the tests and the server node are executed properly.

To test this a new test is added to the server spec to test a correct
close.